backend/requirement.txt
alembic==1.14.0
aniso8601==9.0.1
APScheduler==3.10.4
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.1.2
WTForms==3.2.1
xlrd==2.0.1
xlwings==0.33.3
